+++ tests/apps/good_flow_app/migrations/0024_add_index_with_condition.py
@@ -12,6 +12,6 @@ class Migration(migrations.Migration):
     operations = [
         migrations.AddIndex(
             model_name='testtable',
-            index=models.Index(condition=models.Q(test_field__isnull=False), fields=['test_field'], name='test_index'),
+            index=models.Index(condition=models.Q(test_field_int__isnull=False), fields=['test_field_int'], name='test_index'),
         ),
     ]