+++ tests/apps/good_flow_app/models.py
@@ -1,8 +1,11 @@
 from django.db import models
+from django.contrib.postgres.search import SearchVectorField
 
 
 class TestTable(models.Model):
-    test_field = models.IntegerField()
+    test_field_int = models.IntegerField()
+    test_field_str = models.CharField(max_length=10)
+    test_field_tsv = SearchVectorField()
 
 
 class RelatedTestTable(models.Model):