+++ tests/unit/test_schema.py
@@ -16,6 +16,15 @@ from django_zero_downtime_migrations.backends.postgres.schema import (
 )
 from tests import skip_for_default_django_backend
 
+if django.VERSION[:2] < (2, 2):
+    from django.contrib.postgres.indexes import (
+        BrinIndex, GinIndex, GistIndex
+    )
+else:
+    from django.contrib.postgres.indexes import (
+        BrinIndex, BTreeIndex, GinIndex, GistIndex, HashIndex, SpGistIndex
+    )
+
 pytestmark = skip_for_default_django_backend
 
 DatabaseSchemaEditor = import_string(settings.DATABASES['default']['ENGINE'] + '.schema.DatabaseSchemaEditor')
@@ -1804,4 +1813,160 @@ def test_drop_meta_index__ok():
     ]
     assert editor.django_sql == [
         'DROP INDEX IF EXISTS "tests_model_field1_9b60dc_idx";',
+    ]
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
+def test_add_meta_brin_index__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, BrinIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == [
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING brin ("field1");',
+    ]
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING brin ("field1");',
+    ]
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_meta_brin_index__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, BrinIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == flexible_statement_timeout(
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING brin ("field1");',
+    )
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING brin ("field1");',
+    ]
+
+
+@pytest.mark.skipif(django.VERSION[:2] < (2, 2), reason='functionality provided in django 2.2')
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
+def test_add_meta_btree_index__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, BTreeIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == [
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING btree ("field1");',
+    ]
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING btree ("field1");',
+    ]
+
+
+@pytest.mark.skipif(django.VERSION[:2] < (2, 2), reason='functionality provided in django 2.2')
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_meta_btree_index__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, BTreeIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == flexible_statement_timeout(
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING btree ("field1");',
+    )
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING btree ("field1");',
+    ]
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
+def test_add_meta_gin_index__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, GinIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == [
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING gin ("field1");',
+    ]
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING gin ("field1");',
+    ]
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_meta_gin_index__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, GinIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == flexible_statement_timeout(
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING gin ("field1");',
+    )
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING gin ("field1");',
+    ]
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
+def test_add_meta_gist_index__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, GistIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == [
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING gist ("field1");',
+    ]
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING gist ("field1");',
+    ]
+
+
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_meta_gist_index__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, GistIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == flexible_statement_timeout(
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING gist ("field1");',
+    )
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING gist ("field1");',
+    ]
+
+
+@pytest.mark.skipif(django.VERSION[:2] < (2, 2), reason='functionality provided in django 2.2')
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
+def test_add_meta_hash_index__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, HashIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == [
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING hash ("field1");',
+    ]
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING hash ("field1");',
+    ]
+
+
+@pytest.mark.skipif(django.VERSION[:2] < (2, 2), reason='functionality provided in django 2.2')
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_meta_hash_index__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, HashIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == flexible_statement_timeout(
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING hash ("field1");',
+    )
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING hash ("field1");',
+    ]
+
+
+@pytest.mark.skipif(django.VERSION[:2] < (2, 2), reason='functionality provided in django 2.2')
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True)
+def test_add_meta_spgist_index__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, SpGistIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == [
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING spgist ("field1");',
+    ]
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING spgist ("field1");',
+    ]
+
+
+@pytest.mark.skipif(django.VERSION[:2] < (2, 2), reason='functionality provided in django 2.2')
+@override_settings(ZERO_DOWNTIME_MIGRATIONS_RAISE_FOR_UNSAFE=True,
+                   ZERO_DOWNTIME_MIGRATIONS_FLEXIBLE_STATEMENT_TIMEOUT=True)
+def test_add_meta_spgist_index__with_flexible_timeout__ok():
+    with cmp_schema_editor() as editor:
+        editor.add_index(Model, SpGistIndex(fields=['field1'], name='tests_model_field1_9b60dc_idx'))
+    assert editor.collected_sql == flexible_statement_timeout(
+        'CREATE INDEX CONCURRENTLY "tests_model_field1_9b60dc_idx" ON "tests_model" USING spgist ("field1");',
+    )
+    assert editor.django_sql == [
+        'CREATE INDEX "tests_model_field1_9b60dc_idx" ON "tests_model" USING spgist ("field1");',
     ]